Add maxent_pos_tagger()
